DECLARATION OF INACTIVITY @jimmyhmiller
Due to lack of participation and communication, I propose this declaration of inactivity on @jimmyhmiller which will remain for one (1) business day. During which time, any comment, vote, proposal, or other activity shall render this declaration discarded. If, however, @jimmyhmiller does not respond within this proposal's lifespan and no other Active Player speaks on his behalf, he will be considered removed from the active player listing and become an Inactive Player.

To reclaim an Active Player status, the subject must submit a request to move from the Inactive Player list onto the Active Player list at which time, the request will be processed, and the subject will again be an Active Player.