327 Village Institutions
This is designed to begin a construct for future, purposeful gameplay. As we progress, I envision villages growing into towns, cities, metropolis, countries, planets, etc... Which must gather resources and grow in population and productivity. The "reset" clause is designed to provide incentive to ensure the safety and prosperity of the people.