330 - Feed The People
This rule governs Village Population's Hunger. The base idea is all hunger, farm production, and starvation calculations will occur once per week. The hunger branch of the nomic-bot is currently ready to automate hunger, farm production, and starvation. Purchasing farms (including SECTION 5) is still a manual process, but will soon be addressed. 

We can discuss the best day of the week for these calculations to happen.