331 - Population Growth
For obvious reasons, we need a way for populations to expand. I am not married to this equation, but it is based on a standard growth model. It uses a base 1.1% growth with adjustments based on hunger / surplus where there is a sweet spot around having a surplus equal to 100% of demand.