Convert code to use es6 modules.
All of this was done mechanically using jscodeshift. I found a few codemods for it and wrote a simple one myself. The only thing that was done manually was in universe because I crossed the es6/commonjs boundry. In other words, this should all work. In the future I plan to start adding tests so we will have more confidence it all works.