Convert objects to named exports.
This one is a bit more iffy. Most of this was mechanical with a codemod I wrote (my first big one!). But I did do some manual things. Overall, I think it will work, but I really do need to have a test suite setup to be sure. But adding tests adds to the overhead of people contributing.

I do think that overall this change is positive even if in inadvertently breaks something. Breaking out these objects in the parts allows for greater flexibility with refactoring.

The next codemod I want to write will get rid of the "import * as" and change it to destructuring. This shouldn't be too hard and will definitely help clean things up.

After that I want to add a linter, experiment with different ways of adding commands, and work more on the framework and documentation for authoring commands.