Made adding a command simpler.
The whole idea here is that commands will be able to live in their own file. You can use the addCommand function to register your command. The only other thing you have to do is import your command in the index.

I'm thinking about doing similar things for the interpreter and building the world.