Moved open to new command style.
This was mostly a test to see what it would take. I actually think it was fairly simple. Even though I had to add things to the world, and add a new command, nothing was difficult about the translation. As time goes on, I will translate more commands over to the new architecture.

One thing I didn't handle in my translation is a proper message for inactivePlayers. That is because I plan on implementing some sort of validation functionality soon. That way those sorts of concerns can be dealt with across commands.