Vastly simplified happy.
I hated the way happy looked. It was so large and really got away from the essence of what it was supposed to do. Plus, I am going to have to implement `/train` and it is also a throttle command. In fact, I think throttled commands will become more common. So, I've abstracted the throttledCommand away. If you want to do a throttled command it becomes incredibly easy. Also, I hated that I had to think about undefined, NaN, and the range of a number when adding. So I made an add function that handles all of those cases for me.

This bot is some of the most fun I've had programming in a long time. Eliminating incidental complexity and allowing yourself the ability to only specify exactly what your problem requires is so satisfying.